Modified local storage

+++ QChat/bin/Debug/AppX/js/main.js
@@ -1,6 +1,6 @@
 ﻿// 在此放置代码!
 $(function () {
-    if (!window.localStorage && !Windows.Storage) {
+    if (!window.localStorage) {
         alert("你的浏览器不支持本地存储，这意味着你做出的所有选择和游戏进度都不会被保存");
     }
     var buttons = $(".button");
@@ -10,8 +10,9 @@ $(function () {
                 window.location.href = "ep1/qchat.html";
                 break;
             case 1:
-                window.localStorage.clear();
-                Windows.Storage.ApplicationData.current.localSettings.values.clear();
+                if (window.confirm("你确定要清空本地存档吗")) {
+                    window.localStorage.clear();
+                }
                 break;
             default:
                 throw RangeError("Unbind button clicked");
